paseo programado, se modifica valor y paseos vencidos desactivados

+++ app/src/main/java/com/example/myapplication/repository/PaseadorRepository.kt
@@ -99,7 +99,7 @@ class PaseadorRepository() {
         if (snapshot.exists()) {
           for (userSnapshot in snapshot.children) {
 
-            userSnapshot.ref.child("tarifa").setValue(newTarifa)
+            userSnapshot.ref.child("tarifa").setValue(newTarifa.toInt())
 
           }
         }
